requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
pillow
seaborn
moviepy
scikit-image
av
opencv-contrib-python
faiss